dashboard/auth/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
pydantic[email]==2.6.4
pydantic-settings==2.2.1
python-jose==3.3.0
cryptography==42.0.5
passlib==1.7.4
bcrypt==3.2.2
email-validator==2.1.1
